refactor: extracts read-only controllers from CRUL controller
